Added CancanConcern in to to encapsulate cancan InlineFormsController methods
This way the functionalities added by inline_forms are mor flexible and easy to integrate into main app controller. For example, this way InlineFormsController functionality can added to others main app Controllers - extending from InlineFormsController - as a whole but also just a part of the functionality by including the concern
